requirements.txt
pycryptodome
psutil
tabulate
